Creating Home recording view #1: Redesigned the HTML and SCSS styling of the home page. (page not yet working on its own, be aware)
NB: This commit only introduces the markup and styling structure! The business logic that the double curly braces
rely on is still missing ! Therefore, many errors will appear if only this stage of the page is retrieved/checked-out.
Please, if using git checkout, use later commits if expecting working results.