Creating Home recording view #8: Made a pauseRecording module.
NOTE: It indeed pauses the audio recorder object but it actually stops Speech-to-text service just like a stopRecording will.
This is because, contrary to audio capturing API, the S2T API doesn't have a "pause" method
and can be freely stopped and then started again without disrupting the recording or the speech recognition.