(MAJOR)  Created a complete  RecordingList view that shows all the recordings the user has created.
NB. There's a small bug that causes the list to not refresh with the newest recording when the app redirects the user to this view directly after creating a new recording.
I'm not sure if the bug exists here or at the level of the SaveRecordingDialog or Home, but it's important it's somewhere in those 3 files.